+++ apps/client/src/App.tsx
@@ -1,5 +1,5 @@
 import { Outlet } from 'react-router-dom';
-import Header from '@components/common/Header';
+import Header from '@components/Header';
 import './App.css';
 
 function App() {
