+++ apps/client/src/components/ErrorCharacter/index.tsx
