+++ apps/client/src/components/LoadingCharacter/index.tsx
