+++ apps/client/src/hooks/useProducer.ts
@@ -2,7 +2,7 @@ import { useEffect, useRef, useState } from 'react';
 import { Transport, Device } from 'mediasoup-client/lib/types';
 import { ConnectTransportResponse, TransportInfo } from '../types/mediasoupTypes';
 import { Socket } from 'socket.io-client';
-import { checkDependencies } from '@/lib/utils';
+import { checkDependencies } from '@/utils/utils';
 
 interface UseProducerProps {
   socket: Socket | null;
