+++ apps/client/src/pages/Broadcast/index.tsx
@@ -1,12 +1,12 @@
-import BroadcastTitle from '@/components/BroadcastTitle';
-import ChatContainer from '@/components/ChatContainer';
-import ErrorCharacter from '@/components/common/ErrorCharacter';
-import { useMediaControls } from '@/hooks/useMediaControls';
-import { useMediaStream } from '@/hooks/useMediaStream';
-import { useProducer } from '@/hooks/useProducer';
-import { useRoom } from '@/hooks/useRoom';
-import { useSocket } from '@/hooks/useSocket';
-import { useTransport } from '@/hooks/useTransport';
+import BroadcastTitle from './BroadcastTitle';
+import ChatContainer from '@components/ChatContainer';
+import ErrorCharacter from '@components/ErrorCharacter';
+import { useMediaControls } from '@hooks/useMediaControls';
+import { useMediaStream } from '@hooks/useMediaStream';
+import { useProducer } from '@hooks/useProducer';
+import { useRoom } from '@hooks/useRoom';
+import { useSocket } from '@hooks/useSocket';
+import { useTransport } from '@hooks/useTransport';
 import ChatIcon from '@components/icons/ChatIcon';
 import MicrophoneOffIcon from '@components/icons/MicrophoneOffIcon';
 import MicrophoneOnIcon from '@components/icons/MicrophoneOnIcon';
